Statusを必要な桁数に応じて型を変えれるよう変更

+++ Sources/PokemonKit/Status/BaseStats.swift
@@ -10,20 +10,20 @@ import Foundation
 extension Pokemon.BaseInfo {
     /// 種族値
     public struct BaseStats: Status {
-        public let hitPoints: UInt16
-        public let attack: UInt16
-        public let defense: UInt16
-        public let specialAttack: UInt16
-        public let specialDefense: UInt16
-        public let speed: UInt16
+        public let hitPoints: UInt8
+        public let attack: UInt8
+        public let defense: UInt8
+        public let specialAttack: UInt8
+        public let specialDefense: UInt8
+        public let speed: UInt8
         
         public init(
-            hitPoints: UInt16,
-            attack: UInt16,
-            defense: UInt16,
-            specialAttack: UInt16,
-            specialDefense: UInt16,
-            speed: UInt16
+            hitPoints: UInt8,
+            attack: UInt8,
+            defense: UInt8,
+            specialAttack: UInt8,
+            specialDefense: UInt8,
+            speed: UInt8
         ) {
             self.hitPoints = hitPoints
             self.attack = attack
